requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
flask-ngrok @ git+https://github.com/gstaff/flask-ngrok@33a2e70898c4fb72b645715dacd8789769a99350
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.19.0
requests==2.26.0
urllib3==1.26.6
Werkzeug==2.0.1
